+++ packages/sdk/src/auth/types.ts
@@ -11,7 +11,7 @@ export interface OauthConfig {
 export interface Storage {
   setItem(key: string, value: string): void | Promise<void>
   getItem(key: string): string | undefined | Promise<string | undefined>
-  clear(): void | Promise<void>
+  removeItem(key: string): void | Promise<void>
 }
 
 export type AuthState = auth.types.RequestTokenResponseData & {